move "s3ify" up the call stack
