+++ clientmanager/clientmanager.py
@@ -53,7 +53,8 @@ def main() -> None:
                     args.memory,
                     args.accounting_group,
                     args.singularity_image,
-                    args.client_startup_json,
+                    # put client_startup_json in S3 bucket
+                    utils.s3ify(args.client_startup_json),
                     args.dryrun,
                 )
                 # report to SkyDriver
