+++ clientmanager/utils.py
@@ -1,7 +1,12 @@
 """General Utilities."""
 
 
+import dataclasses as dc
+from pathlib import Path
+
+import boto3  # type: ignore[import]
 import htcondor  # type: ignore[import]
+import requests
 from rest_tools.client import RestClient
 
 from .config import ENV, LOGGER
@@ -42,3 +47,56 @@ def update_skydriver(
             }
         },
     )
+
+
+boto3.set_stream_logger(name="botocore")
+
+
+@dc.dataclass
+class S3File:
+    """Wrap an S3 file."""
+
+    url: str
+    base_fname: str
+
+
+def s3ify(filepath: Path) -> S3File:
+    """Put the file in s3 and return info about it."""
+    if not (
+        ENV.EWMS_TMS_S3_URL
+        and ENV.EWMS_TMS_S3_ACCESS_KEY
+        and ENV.EWMS_TMS_S3_SECRET_KEY
+        and ENV.EWMS_TMS_S3_BUCKET
+        and ENV.SKYSCAN_SKYDRIVER_SCAN_ID
+    ):
+        raise RuntimeError(
+            "must define all EWMS_TMS_S3_* environment variables to use S3"
+        )
+    s3_client = boto3.client(
+        "s3",
+        "us-east-1",
+        endpoint_url=ENV.EWMS_TMS_S3_URL,
+        aws_access_key_id=ENV.EWMS_TMS_S3_ACCESS_KEY,
+        aws_secret_access_key=ENV.EWMS_TMS_S3_SECRET_KEY,
+    )
+    bucket = ENV.EWMS_TMS_S3_BUCKET
+    key = ENV.SKYSCAN_SKYDRIVER_SCAN_ID
+
+    # POST
+    upload_details = s3_client.generate_presigned_post(bucket, key)
+    with open(filepath, "rb") as f:
+        response = requests.post(
+            upload_details["url"],
+            data=upload_details["fields"],
+            files={"file": (filepath.name, f)},  # maps filename to obj
+        )
+    LOGGER.info(f"Upload response: {response.status_code}")
+    LOGGER.info(str(response.content))
+
+    # get GET url
+    get_url = s3_client.generate_presigned_url(
+        "get_object",
+        Params={"Bucket": bucket, "Key": key},
+        ExpiresIn=3600,  # 60 mins
+    )
+    return S3File(get_url, filepath.name)
